Move listener to correct package
